requirements.txt
fastapi
uvicorn
python-dotenv
feedparser